requirements.txt
flask==2.0.1
pandas==1.3.3
google-generativeai==0.3.1
python-dotenv==0.19.1
pytest==6.2.5
gunicorn==20.1.0 